fix tests, move in another package + refacto given, fix name
